requirements.txt
kaldiio==2.18.0
librosa==0.9.1
matplotlib==3.7.1
multiprocess==0.70.16
numpy==1.24.4
pathos==0.3.2
pip==23.2.1
pytest==8.1.1
pyworld==0.3.4
PyYAML==6.0
resampy==0.4.3
Resemblyzer==0.1.3
s3prl==0.4.15
scikit_learn==1.3.0
scipy==1.13.0
setuptools==68.0.0
soundfile==0.12.1
tensorboardX==2.6.2.2
torch==1.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.56.0
webrtcvad==2.0.10
